改了下结构，但是报错writeshortest未定义，但是我找不到这个writeshortest
