+++ Project.toml
@@ -1,1 +1,3 @@
 [deps]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
